requirements.txt
python-dotenv==1.0.1
prometheus-client==0.21.0
flask==3.0.3
flask-restx==1.3.0
pycouchdb==1.16.0